Move notify into top-level utils package
